Fixed duplicate variable names

+++ lib/MC/MCAsmStreamer.cpp
@@ -673,30 +673,30 @@ void MCAsmStreamer::EmitValueImpl(const MCExpr *Value, unsigned Size) {
     Directive = MAI->getData32bitsDirective();
     // If the target doesn't support 32-bit data, emit as two 16-bit halves.
     if (Directive) break;
-    int32_t IntValue;
-    if (!Value->EvaluateAsAbsolute(IntValue))
+    int32_t Int32Value;
+    if (!Value->EvaluateAsAbsolute(Int32Value))
       report_fatal_error("Don't know how to emit this value.");
     if (MAI->isLittleEndian()) {
-      EmitIntValue((uint16_t)(IntValue >> 0 ), 2);
-      EmitIntValue((uint16_t)(IntValue >> 16), 2);
+      EmitIntValue((uint16_t)(Int32Value >> 0 ), 2);
+      EmitIntValue((uint16_t)(Int32Value >> 16), 2);
     } else {
-      EmitIntValue((uint16_t)(IntValue >> 16), 2);
-      EmitIntValue((uint16_t)(IntValue >> 0 ), 2);
+      EmitIntValue((uint16_t)(Int32Value >> 16), 2);
+      EmitIntValue((uint16_t)(Int32Value >> 0 ), 2);
     }
     return;
   case 8:
     Directive = MAI->getData64bitsDirective();
     // If the target doesn't support 64-bit data, emit as two 32-bit halves.
     if (Directive) break;
-    int64_t IntValue;
-    if (!Value->EvaluateAsAbsolute(IntValue))
+    int64_t Int64Value;
+    if (!Value->EvaluateAsAbsolute(Int64Value))
       report_fatal_error("Don't know how to emit this value.");
     if (MAI->isLittleEndian()) {
-      EmitIntValue((uint32_t)(IntValue >> 0 ), 4);
-      EmitIntValue((uint32_t)(IntValue >> 32), 4);
+      EmitIntValue((uint32_t)(Int64Value >> 0 ), 4);
+      EmitIntValue((uint32_t)(Int64Value >> 32), 4);
     } else {
-      EmitIntValue((uint32_t)(IntValue >> 32), 4);
-      EmitIntValue((uint32_t)(IntValue >> 0 ), 4);
+      EmitIntValue((uint32_t)(Int64Value >> 32), 4);
+      EmitIntValue((uint32_t)(Int64Value >> 0 ), 4);
     }
     return;
   }
